+++ tests/test_action.py
@@ -1,4 +1,4 @@
-"""Tests for action data points of hahomematic."""
+"""Tests for action data points of aiohomematic."""
 
 from __future__ import annotations
 
@@ -7,10 +7,10 @@ from unittest.mock import Mock, call
 
 import pytest
 
-from hahomematic.central import CentralUnit
-from hahomematic.client import Client
-from hahomematic.const import DataPointUsage, ParamsetKey
-from hahomematic.model.generic import DpAction
+from aiohomematic.central import CentralUnit
+from aiohomematic.client import Client
+from aiohomematic.const import DataPointUsage, ParamsetKey
+from aiohomematic.model.generic import DpAction
 
 from tests import helper