+++ tests/test_binary_sensor.py
@@ -1,4 +1,4 @@
-"""Tests for binary_sensor data points of hahomematic."""
+"""Tests for binary_sensor data points of aiohomematic."""
 
 from __future__ import annotations
 
@@ -8,11 +8,11 @@ from unittest.mock import Mock
 
 import pytest
 
-from hahomematic.central import CentralUnit
-from hahomematic.client import Client
-from hahomematic.const import DataPointUsage
-from hahomematic.model.generic import DpBinarySensor
-from hahomematic.model.hub import SysvarDpBinarySensor
+from aiohomematic.central import CentralUnit
+from aiohomematic.client import Client
+from aiohomematic.const import DataPointUsage
+from aiohomematic.model.generic import DpBinarySensor
+from aiohomematic.model.hub import SysvarDpBinarySensor
 
 from tests import const, helper