+++ tests/test_cover.py
@@ -1,4 +1,4 @@
-"""Tests for cover data points of hahomematic."""
+"""Tests for cover data points of aiohomematic."""
 
 from __future__ import annotations
 
@@ -8,11 +8,11 @@ from unittest.mock import DEFAULT, Mock, call
 
 import pytest
 
-from hahomematic.central import CentralUnit
-from hahomematic.client import Client
-from hahomematic.const import WAIT_FOR_CALLBACK, DataPointUsage, ParamsetKey
-from hahomematic.model.custom import CustomDpBlind, CustomDpCover, CustomDpGarage, CustomDpIpBlind, CustomDpWindowDrive
-from hahomematic.model.custom.cover import (
+from aiohomematic.central import CentralUnit
+from aiohomematic.client import Client
+from aiohomematic.const import WAIT_FOR_CALLBACK, DataPointUsage, ParamsetKey
+from aiohomematic.model.custom import CustomDpBlind, CustomDpCover, CustomDpGarage, CustomDpIpBlind, CustomDpWindowDrive
+from aiohomematic.model.custom.cover import (
     _CLOSED_LEVEL,
     _OPEN_LEVEL,
     _OPEN_TILT_LEVEL,