+++ tests/test_event.py
@@ -1,4 +1,4 @@
-"""Tests for switch data points of hahomematic."""
+"""Tests for switch data points of aiohomematic."""
 
 from __future__ import annotations
 
@@ -7,10 +7,10 @@ from unittest.mock import Mock, call
 
 import pytest
 
-from hahomematic.central import CentralUnit
-from hahomematic.client import Client
-from hahomematic.const import DataPointUsage, EventType
-from hahomematic.model.event import ClickEvent, DeviceErrorEvent, ImpulseEvent
+from aiohomematic.central import CentralUnit
+from aiohomematic.client import Client
+from aiohomematic.const import DataPointUsage, EventType
+from aiohomematic.model.event import ClickEvent, DeviceErrorEvent, ImpulseEvent
 
 from tests import const, helper