+++ tests/test_light.py
@@ -1,4 +1,4 @@
-"""Tests for light data points of hahomematic."""
+"""Tests for light data points of aiohomematic."""
 
 from __future__ import annotations
 
@@ -7,10 +7,10 @@ from unittest.mock import Mock, call
 
 import pytest
 
-from hahomematic.central import CentralUnit
-from hahomematic.client import Client
-from hahomematic.const import WAIT_FOR_CALLBACK, DataPointUsage, ParamsetKey
-from hahomematic.model.custom import (
+from aiohomematic.central import CentralUnit
+from aiohomematic.client import Client
+from aiohomematic.const import WAIT_FOR_CALLBACK, DataPointUsage, ParamsetKey
+from aiohomematic.model.custom import (
     CustomDpColorDimmer,
     CustomDpColorDimmerEffect,
     CustomDpColorTempDimmer,
@@ -18,7 +18,7 @@ from hahomematic.model.custom import (
     CustomDpIpFixedColorLight,
     CustomDpIpRGBWLight,
 )
-from hahomematic.model.custom.light import _NOT_USED, _ColorBehaviour, _FixedColor, _TimeUnit
+from aiohomematic.model.custom.light import _NOT_USED, _ColorBehaviour, _FixedColor, _TimeUnit
 
 from tests import const, helper