+++ tests/test_switch.py
@@ -1,4 +1,4 @@
-"""Tests for switch data points of hahomematic."""
+"""Tests for switch data points of aiohomematic."""
 
 from __future__ import annotations
 
@@ -7,12 +7,12 @@ from unittest.mock import Mock, call
 
 import pytest
 
-from hahomematic.central import CentralUnit
-from hahomematic.client import Client
-from hahomematic.const import WAIT_FOR_CALLBACK, DataPointUsage, ParamsetKey
-from hahomematic.model.custom import CustomDpSwitch
-from hahomematic.model.generic import DpSwitch
-from hahomematic.model.hub import SysvarDpSwitch
+from aiohomematic.central import CentralUnit
+from aiohomematic.client import Client
+from aiohomematic.const import WAIT_FOR_CALLBACK, DataPointUsage, ParamsetKey
+from aiohomematic.model.custom import CustomDpSwitch
+from aiohomematic.model.generic import DpSwitch
+from aiohomematic.model.hub import SysvarDpSwitch
 
 from tests import helper